requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
OdooRPC==0.10.1
packaging==23.2
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
python-dotenv==1.0.0
typing_extensions==4.8.0
watchdog==3.0.0
Werkzeug==3.0.1
